\name{glmmlasso-package}
\alias{glmmlasso-package}
\docType{package}
\title{glmmlasso}
\description{Fits generalized linear mixed models with a Lasso penalty
  for the fixed effects.}
\details{
\tabular{ll}{
Package: \tab glmmlasso\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-09-13\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

This is the first version of the package and subject to testing.
}
\author{
Juerg Schelldorfer

Maintainer: Juerg Schelldorfer <schell@stat.math.ethz.ch>
}

\keyword{package}
